+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Function `pad_and_crop_dim` and class `PadCropDim` to pa then crop a tensor along a specified dimension.
 - Function `reimport_modules` to re-import modules and functions in a notebook.
 - Function `checksum_module` to compare module weights.
+- Function `nelement` to get number of elements in Tensor-like object.
 
 ### Changed
 - Rename `is_pickle_root` to `is_packed_root` (old name was kept for backward compatibility).